OLS-1229: HuggingFace environment settings
