+++ ols/utils/environments.py
@@ -3,6 +3,8 @@
 import os
 import tempfile
 
+import ols.app.models.config as config_model
+
 
 def configure_gradio_ui_envs() -> None:
     """Configure GradioUI framework environment variables."""
@@ -15,3 +17,17 @@ def configure_gradio_ui_envs() -> None:
     # Fixes: https://issues.redhat.com/browse/OLS-301
     tempdir = os.path.join(tempfile.gettempdir(), "matplotlib")
     os.environ["MPLCONFIGDIR"] = tempdir
+
+
+def configure_hugging_face_envs(ols_config: config_model.OLSConfig) -> None:
+    """Configure HuggingFace library environment variables."""
+    if (
+        ols_config
+        and hasattr(ols_config, "reference_content")
+        and hasattr(ols_config.reference_content, "embeddings_model_path")
+        and ols_config.reference_content.embeddings_model_path
+    ):
+        os.environ["TRANSFORMERS_CACHE"] = str(
+            ols_config.reference_content.embeddings_model_path
+        )
+        os.environ["TRANSFORMERS_OFFLINE"] = "1"
